web3-modal: allow custom connect function
