+++ packages/web3-modal/src/hooks/index.ts
@@ -5,3 +5,4 @@ export * from './useMergeThemes'
 export * from './usePstlWeb3Modal'
 export * from './usePstlWeb3ModalStore'
 export * from './useLimitChainsAndSwitchCallback'
+export * from './useAllWeb3Modals'
