+++ packages/web3-modal/src/hooks/useAllWeb3Modals.ts
@@ -1,0 +1,13 @@
+import { useWeb3Modal } from "@web3modal/react"
+import { usePstlWeb3Modal } from "./usePstlWeb3Modal"
+
+export type AllWeb3ModalStore = {
+  root: ReturnType<typeof usePstlWeb3Modal>
+  walletConnect: ReturnType<typeof useWeb3Modal>
+}
+export function useAllWeb3Modals(): AllWeb3ModalStore {
+    return {
+      root: usePstlWeb3Modal(),
+      walletConnect: useWeb3Modal()
+    }
+  }
