+++ packages/web3-modal/src/utils/misc.ts
@@ -1,0 +1,5 @@
+export function trimAndLowerCase(thing: string | undefined) {
+  if (!thing) return ''
+
+  return thing.replace(' ', '').toLowerCase()
+}
